[authorizer] Option to keep authorized hosts on exit
* [firewall/nft] update FlushTable and DeleteChain
* [firewall/nft] rename chain.go to chain4.go
* [nettrust] move makeDefaults before authorizer is called
  this is to ensure that authorizer will be able to read
  existing sets
* [nettrust] suffix chain and table variables with respected hook
* [nettrust] delete also input chain from the net-trust table
* [readme.md] update with new NetTrust options

+++ authorizer/activity.go
@@ -50,13 +50,15 @@ func (f *Authorizer) ttlCacheChecker() (*ServiceContext, error) {
 					"Stage":     "Deauthorize",
 				})
 
-				for h := range f.cache.Hosts {
-					l.Infof("Removing host [%s] from firewall rules", h)
-					err := f.fw.DeleteIPv4FromSetRule(f.authorizedSet, h)
-					if err != nil {
-						l.Error(err)
+				if !f.doNotFlushAuthorizedHosts {
+					for h := range f.cache.Hosts {
+						l.Infof("Removing host [%s] from firewall rules", h)
+						err := f.fw.DeleteIPv4FromSetRule(f.authorizedSet, h)
+						if err != nil {
+							l.Error(err)
+						}
+						f.cache.Delete(h)
 					}
-					f.cache.Delete(h)
 				}
 
 				l.Info("Bye!")
